+++ error_messages_test.ts
@@ -19,7 +19,13 @@ test({
       typeof mod.ERROR_MISSING_STATIC_OR_VALUE_ON_POLYMORPHIC_SWITCH,
       "string",
     );
+    assertEquals(typeof mod.ERROR_INVALID_DATE, "string");
+    assertEquals(typeof mod.ERROR_TO_OBJECT_CONTAINING_INVALID_VALUE, "string");
+    assertEquals(
+      typeof mod.ERROR_TO_OBJECT_CONTAINING_INVALID_SUB_VALUE,
+      "string",
+    );
 
-    assertEquals(Object.keys(mod).length, 6);
+    assertEquals(Object.keys(mod).length, 9);
   },
 });