+++ mod.ts
@@ -17,6 +17,8 @@ export type {
 } from "./strategy/compose_strategy.ts";
 /** from json strategies */
 export { toSerializable } from "./strategy/from_json/to_serializable.ts";
+export { toObjectContaining } from "./strategy/from_json/to_object_containing.ts";
+export { fromObjectContaining } from "./strategy/to_json/from_object_containing.ts";
 export { createDateStrategy, iso8601Date } from "./strategy/from_json/date.ts";
 /** polymorphic classes */
 export {