+++ strategy/from_json/date.ts
@@ -1,14 +1,19 @@
 // Copyright 2018-2020 Gamebridge.ai authors. All rights reserved. MIT license.
 import { FromJSONStrategy } from "../compose_strategy.ts";
 import { JSONValue } from "../../serializable.ts";
+import { ERROR_INVALID_DATE } from "../../error_messages.ts";
 
 /** allows authors to pass a regex to parse as a date */
 export function createDateStrategy(regex: RegExp): FromJSONStrategy {
   return function _createDateStrategy(input: JSONValue): Date {
     if (typeof input !== "string" || !regex.exec(input)) {
-      throw new Error("invalid date");
+      throw new Error(ERROR_INVALID_DATE);
     }
-    return new Date(input);
+    const date = new Date(input);
+    if (date.toJSON() === null) {
+      throw new Error(ERROR_INVALID_DATE);
+    }
+    return date;
   };
 }
 
@@ -23,7 +28,7 @@ export function iso8601Date(input: JSONValue): any {
     //        <year>              -    <month>    -         <day>   <T non optional> <hour>:    <minute>:    <second>    <.milli> <Z  or `-+timezone`>
     /^(-?(?:[1-9][0-9]*)?[0-9]{4})-(1[0-2]|0[1-9])-(3[01]|0[1-9]|[12][0-9])T(2[0-3]|[01][0-9]):([0-5][0-9]):([0-5][0-9])(\.[0-9]+)?(Z|[+-](?:2[0-3]|[01][0-9]):[0-5][0-9])?$/;
   if (typeof input !== "string" || !iso.exec(input)) {
-    throw new Error("Invalid date");
+    throw new Error(ERROR_INVALID_DATE);
   }
   return new Date(input);
 }