+++ strategy/from_json/to_serializable.ts
@@ -2,17 +2,18 @@
 
 import { JSONValue, Serializable } from "../../serializable.ts";
 import { FromJSONStrategy } from "../compose_strategy.ts";
+import { getNew, StrategyTypeArgument } from "../_utils.ts";
 
 /** revive data using `fromJSON` on a subclass type */
 export function toSerializable<T>(
-  type: T & { new (): Serializable },
+  type: StrategyTypeArgument<T>,
 ): FromJSONStrategy {
   return function _toSerializable(
     value: JSONValue,
   ): Serializable | Serializable[] {
     if (Array.isArray(value)) {
-      return value.map((item) => new type().fromJSON(item));
+      return value.map((item) => getNew(type).fromJSON(item));
     }
-    return new type().fromJSON(value);
+    return getNew(type).fromJSON(value);
   };
 }