+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- `toObjectContaining` and `fromObjectContaining` fixing #89
+
+## Changed
+- deno@1.6.0
+- std@0.80.0
+
 ## [v1.0.0] - 2020-11-29
 
 ### Added